+++ packages/integration-sdk-runtime/src/execution/__tests__/jobState.test.ts
@@ -88,37 +88,3 @@ describe('#findEntity', () => {
     expect(await jobState.findEntity('invalid-entity-key')).toEqual(null);
   });
 });
-
-describe('#getEntity', () => {
-  afterEach(() => {
-    vol.reset();
-  });
-
-  test('should get entity by _key', async () => {
-    const params = getMockCreateStepJobStateParams();
-    const jobState = createStepJobState(params);
-    const entity: Entity = {
-      _type: 'a_entity',
-      _class: 'A',
-      _key: 'a',
-    };
-
-    await jobState.addEntity(entity);
-    expect(await jobState.getEntity('a')).toStrictEqual(entity);
-  });
-
-  test('should throw if entity does not exist', async () => {
-    expect.assertions(2);
-    const params = getMockCreateStepJobStateParams();
-    const jobState = createStepJobState(params);
-
-    try {
-      await jobState.getEntity('invalid-entity-key');
-    } catch (err) {
-      expect(err instanceof IntegrationMissingKeyError).toEqual(true);
-      expect(err.message).toEqual(
-        `Failed to find entity in in-memory graph object metadata store (_key=invalid-entity-key)`,
-      );
-    }
-  });
-});